adaptive_ai/requirements.txt
# Core dependencies
litserve>=0.1.0
pydantic>=2.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.16.0

# API and web
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# Configuration and utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
typing-extensions>=4.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
ruff>=0.0.284 